Moved livekit_voice_agent.py to seperate folder and created docker container
Livekit voice agent must be run as a seperate process in a docker container and managed using kubernetics
