--- .env.example
@@ -1,43 +1,0 @@
-# Backend Selection
-VITE_USE_PYTHON_BACKEND=true
-VITE_PYTHON_API_URL=http://localhost:3001
-VITE_API_URL=http://localhost:3001/api
-
-# LLM Provider Configuration
-# Choose 'openai', 'anthropic', or 'gemini'
-LLM_PROVIDER=gemini
-
-# OpenAI Configuration
-OPENAI_API_KEY=your_openai_api_key_here
-
-# Anthropic Configuration (alternative to OpenAI)
-# ANTHROPIC_API_KEY=your_anthropic_api_key_here
-
-# Google Gemini Configuration (recommended for Pydantic AI)
-GEMINI_API_KEY=your_gemini_api_key_here
-GOOGLE_API_KEY=your_google_api_key_here
-
-# Google Cloud Configuration (for Speech-to-Text and Text-to-Speech)
-GOOGLE_APPLICATION_CREDENTIALS=path/to/your/service-account-key.json
-GOOGLE_CLOUD_PROJECT_ID=your_google_cloud_project_id
-
-# Voice Agent Provider Configuration
-# Choose 'openai' or 'google' for voice agents
-VOICE_AGENT_PROVIDER=google
-
-# LiveKit Configuration for Voice Interviews
-LIVEKIT_API_KEY=your_livekit_api_key
-LIVEKIT_API_SECRET=your_livekit_api_secret
-LIVEKIT_WS_URL=wss://your-livekit-server.com
-LIVEKIT_URL=wss://your-livekit-server.com
-
-# Server Configuration
-PORT=3001
-ENVIRONMENT=development
-
-# Optional: JWT Secret for additional security
-JWT_SECRET=your_jwt_secret_for_additional_security
-
-# Additional API Keys
-ELEVEN_API_KEY=your_eleven_api_key
-GLADIA_API_KEY=your_gladia_api_key
