+++ python_backend/livekit_voice_agent/requirements.txt
@@ -1,0 +1,26 @@
+# Core backend dependencies
+python-dotenv==1.0.1
+python-multipart==0.0.12
+loguru==0.7.2
+
+# LLM providers - exact compatible versions
+openai==1.84.0
+google-generativeai==0.8.3
+
+# LiveKit core - using latest 1.x versions for compatibility with 1.1.4 plugins
+livekit==1.0.11
+livekit-agents==1.1.7
+livekit-api==1.0.3
+
+# LiveKit dependencies
+transformers >=4.47.1
+numpy >=1.26
+onnxruntime >=1.18
+jinja2
+
+livekit-plugins-gladia==1.1.7        # Latest available version
+livekit-plugins-silero==1.1.7
+livekit-plugins-google==1.1.7
+livekit-plugins-elevenlabs==1.1.7           # Latest available version
+livekit-plugins-turn-detector==1.1.7  # Latest available version
+
