+++ requirements.txt
@@ -1,6 +1,5 @@
 # Core backend dependencies
 fastapi==0.115.6
-uvicorn==0.32.1
 python-dotenv==1.0.1
 httpx==0.28.1
 aiofiles==24.1.0
@@ -8,24 +7,6 @@ python-multipart==0.0.12
 loguru==0.7.2
 websockets==13.1
 
-# Pydantic - exact compatible versions
-pydantic==2.10.3
-pydantic-ai-slim[openai]==0.0.14
-
 # LLM providers - exact compatible versions
 openai==1.84.0
 google-generativeai==0.8.3
-
-# LiveKit core - using latest 1.x versions for compatibility with 1.1.4 plugins
-livekit==1.0.11
-livekit-agents==1.1.4
-livekit-api==1.0.3
-
-# LiveKit plugins - using ACTUAL available versions from error messages
-
-livekit-plugins-gladia==1.1.4        # Latest available version
-livekit-plugins-silero==1.1.4
-livekit-plugins-google
-livekit-plugins-elevenlabs           # Latest available version
-livekit-plugins-turn-detector==1.1.4  # Latest available version
-
